댓글 기능

+++ src/main/java/com/sparta/service/CommentService.java
@@ -15,7 +15,7 @@ import lombok.RequiredArgsConstructor;
 public class CommentService {
 	private final CommentRepository commentRepository;
 	@Transactional
-	public Long update(Long id, CommentRequestDto requestDto) {
+	public int update(int id, CommentRequestDto requestDto) {
 		Comment comment = commentRepository.findById(id).orElseThrow(
 		() -> new IllegalArgumentException("아이디가 존재하지 않습니다.")
 		);
